Migrate live/options/test-timestamp-domain to pytest
